requirements.txt
dnspython==2.1.0
colorama==0.4.4
pyfiglet==0.8.post1